# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Dokumentasi Swagger untuk backend be_latihan menggunakan Golang Fiber, GORM, PostgreSQL, dan JWT Bearer Token.",
        "title": "API be_latihan Praktikum 13",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Praktikum Pemrograman III",
            "email": "praktikum@example.com"
        },
        "version": "1.0"
    },
    "host": "127.0.0.1:3000",
    "basePath": "/",
    "paths": {
...
        "/change-password": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mengubah password user yang sedang login menggunakan token JWT.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Ubah password user",
                "parameters": [
                    {
                        "description": "Data perubahan password",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.ChangePasswordRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Password berhasil diubah",
                        "schema": {
                            "$ref": "#/definitions/model.ChangePasswordSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Payload atau konfirmasi password tidak valid",
                        "schema": {
                            "$ref": "#/definitions/model.BadRequestResponse"
                        }
                    },
                    "401": {
                        "description": "Token tidak valid atau password lama salah",
                        "schema": {
                            "$ref": "#/definitions/model.UnauthorizedResponse"
                        }
                    },
                    "500": {
                        "description": "Gagal mengubah password",
                        "schema": {
                            "$ref": "#/definitions/model.InternalServerErrorResponse"
                        }
                    }
                }
            }
        },
        "/login": {
            "post": {
                "description": "Melakukan autentikasi user dan mengembalikan token JWT untuk akses endpoint yang dilindungi.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Login user",
                "parameters": [
                    {
                        "description": "Data login user",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.AuthRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login berhasil",
                        "schema": {
                            "$ref": "#/definitions/model.LoginSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Payload tidak valid",
                        "schema": {
                            "$ref": "#/definitions/model.BadRequestResponse"
                        }
                    },
                    "401": {
                        "description": "Username atau password salah",
                        "schema": {
                            "$ref": "#/definitions/model.UnauthorizedResponse"
                        }
                    },
                    "500": {
                        "description": "Gagal memproses login",
                        "schema": {
                            "$ref": "#/definitions/model.InternalServerErrorResponse"
                        }
                    }
                }
            }
        },
        "/register": {
            "post": {
                "description": "Membuat akun user baru untuk mendapatkan akses ke API.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Registrasi user baru",
                "parameters": [
                    {
                        "description": "Data registrasi user",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.AuthRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Registrasi berhasil",
                        "schema": {
                            "$ref": "#/definitions/model.RegisterSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Payload tidak valid",
                        "schema": {
                            "$ref": "#/definitions/model.BadRequestResponse"
                        }
                    },
                    "409": {
                        "description": "Username sudah digunakan atau data tidak valid",
                        "schema": {
                            "$ref": "#/definitions/model.ConflictResponse"
                        }
                    },
                    "500": {
                        "description": "Gagal membuat hash password",
                        "schema": {
                            "$ref": "#/definitions/model.InternalServerErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.AuthRequest": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string",
                    "example": "password123"
                },
                "role": {
                    "type": "string",
                    "example": "admin"
                },
                "username": {
                    "type": "string",
                    "example": "admin"
                }
            }
        },
        "model.AuthUserResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "role": {
                    "type": "string",
                    "example": "admin"
                },
                "username": {
                    "type": "string",
                    "example": "admin"
                }
            }
        },
        "model.BadRequestResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "detail error validasi"
                },
                "message": {
                    "type": "string",
                    "example": "payload tidak valid"
                }
            }
        },
        "model.ChangePasswordRequest": {
            "type": "object",
            "properties": {
                "confirm_password": {
                    "type": "string",
                    "example": "passwordBaru123"
                },
                "new_password": {
                    "type": "string",
                    "example": "passwordBaru123"
                },
                "old_password": {
                    "type": "string",
                    "example": "password123"
                }
            }
        },
        "model.ChangePasswordSuccessResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "password berhasil diubah"
                }
            }
        },
        "model.ConflictResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "duplicate key value violates unique constraint"
                },
                "message": {
                    "type": "string",
                    "example": "username sudah digunakan atau data tidak valid"
                }
            }
        },
...
        "model.InternalServerErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "detail error server"
                },
                "message": {
                    "type": "string",
                    "example": "terjadi kesalahan pada server"
                }
            }
        },
        "model.LoginResponse": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.example"
                },
                "user": {
                    "$ref": "#/definitions/model.AuthUserResponse"
                }
            }
        },
        "model.LoginSuccessResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/model.LoginResponse"
                },
                "message": {
                    "type": "string",
                    "example": "login berhasil"
                }
            }
        },
...
        "model.RegisterSuccessResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/model.AuthUserResponse"
                },
                "message": {
                    "type": "string",
                    "example": "register berhasil"
                }
            }
        },
        "model.UnauthorizedResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "authorization header wajib diisi"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Masukkan token dengan format: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}